ボタンコンポーネントの編集とoutlineバージョン作成
